replication-package-RQ3/EcoPipe/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
flask==3.0.3
fonttools==4.51.0
httmock==1.4.0
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.4.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
pillow==10.3.0
prometheus-api-client==0.5.5
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
